Update for publication

+++ infra/gkx/gkx/mdx/interface.py
@@ -26,6 +26,7 @@ defaults = {
     "batch_size": 25,
     "chunk_size": 2500,
     "outfolder": Path("trajectory/"),
+    "maxsteps": 25000,
 }
 
 
@@ -44,6 +45,7 @@ def run_from_dict(dct):
     potential = dct.pop("potential")
     calculator = dct.pop("calculator")
 
+    # todo: make exception for recompute?
     maxsteps = dct.pop("maxsteps")
     batch_size = dct.pop("batch_size")
     chunk_size = dct.pop("chunk_size")
@@ -67,23 +69,30 @@ def run_from_dict(dct):
     else:
         primitive_atoms = None
 
-    outfolder = dct.pop("outfolder")
+    outfolder = Path(dct.pop("outfolder"))
 
     if outfolder.is_dir():
-        if (outfolder / "TMP").is_file():
-            comms.warn("found TMP file; deleting. (did a job shut down hard?)")
-            (outfolder / "TMP").unlink()
+        if "recompute" not in dct:
+            if (outfolder / "TMP").is_file():
+                comms.warn("found TMP file; deleting. (did a job shut down hard?)")
+                (outfolder / "TMP").unlink()
 
-        n_chunks = len(list(outfolder.glob("*.nc")))
-        if n_chunks > 0:
-            traj = open_trajectory(outfolder)
-            initial_atoms = traj.get_atoms(-1)
-            initial_step = len(traj)
-            initial_chunk = n_chunks
-            comms.talk(f"restarting from step {initial_step}, chunk {initial_chunk}")
-        else:
-            initial_step = 0
-            initial_chunk = 0
+            n_chunks = len(list(outfolder.glob("*.nc")))
+            if n_chunks > 0:
+                traj = open_trajectory(outfolder)
+                initial_step = len(traj.time)
+                if initial_step >= maxsteps:
+                    comms.talk("already done, quitting")
+                    return None
+
+                initial_atoms = traj.get_atoms(-1)
+                initial_chunk = n_chunks
+                comms.talk(
+                    f"restarting from step {initial_step}, chunk {initial_chunk}"
+                )
+            else:
+                initial_step = 0
+                initial_chunk = 0
 
     else:
         outfolder.mkdir()
@@ -110,20 +119,14 @@ def run_from_dict(dct):
             initial_chunk=initial_chunk,
         )
 
-    elif "nvt" in dct:
-        from .nvt import run
-
-        dt = dct["nvt"]["dt"]
-        friction = dct["nvt"]["friction"]
-        temperature = dct["nvt"]["temperature"]
+    elif "nvt" in dct or "npt" in dct:
+        from .nvt_npt import run
 
         run(
             maxsteps,
             potential,
             calculator,
-            dt,
-            temperature,
-            friction,
+            dct,
             initial_atoms,
             supercell_atoms=supercell_atoms,
             primitive_atoms=primitive_atoms,
@@ -131,4 +134,52 @@ def run_from_dict(dct):
             chunk_size=chunk_size,
             initial_step=initial_step,
             initial_chunk=initial_chunk,
+        )
+
+    elif "recompute" in dct:
+        from .recompute import run
+
+        assert str(dct["recompute"]["trajectory"]) != str(outfolder)
+
+        trajectory = open_trajectory(dct["recompute"]["trajectory"])
+        step_size = dct["recompute"]["step_size"]
+        devices = dct["recompute"].get("devices", 1)
+
+        if outfolder.is_dir():
+            if (outfolder / "TMP").is_file():
+                comms.warn("found TMP file; deleting. (did a job shut down hard?)")
+                (outfolder / "TMP").unlink()
+
+            n_chunks = len(list(outfolder.glob("*.nc")))
+            if n_chunks > 0:
+                traj = open_trajectory(outfolder)
+                initial_step = len(traj)
+
+                if (initial_step * step_size) >= len(trajectory.time):
+                    comms.talk("already done, quitting")
+                    return None
+
+                initial_chunk = n_chunks
+                initial_atoms = trajectory.get_atoms(initial_step * step_size)
+                comms.talk(
+                    f"restarting from (recompute) step {initial_step}, chunk {initial_chunk}"
+                )
+            else:
+                initial_step = 0
+                initial_chunk = 0
+
+        run(
+            trajectory,
+            potential,
+            calculator,
+            step_size,
+            initial_atoms,
+            supercell_atoms=supercell_atoms,
+            primitive_atoms=primitive_atoms,
+            outfolder=outfolder,
+            chunk_size=chunk_size,
+            batch_size=batch_size,
+            initial_step=initial_step,
+            initial_chunk=initial_chunk,
+            devices=devices,
         )
